+++ src/Form/Type/DatePickerType.php
@@ -27,6 +27,8 @@ final class DatePickerType extends BasePickerType
             'dp_pick_time' => false,
             'format' => DateType::DEFAULT_FORMAT,
         ]));
+
+        parent::configureOptions($resolver);
     }
 
     public function getParent(): string