+++ src/Form/Type/DateTimePickerType.php
@@ -30,6 +30,8 @@ final class DateTimePickerType extends BasePickerType
             'format' => DateTimeType::DEFAULT_DATE_FORMAT,
             'date_format' => null,
         ]));
+
+        parent::configureOptions($resolver);
     }
 
     public function getParent(): string