+++ src/Model/Adapter/AdapterChain.php
@@ -13,38 +13,15 @@ declare(strict_types=1);
 
 namespace Sonata\CoreBundle\Model\Adapter;
 
-final class AdapterChain implements AdapterInterface
+@trigger_error(
+    'The '.__NAMESPACE__.'\AdapterInterface class is deprecated since version 3.x and will be removed in 4.0.'
+    .' Use Sonata\Doctrine\Adapter\AdapterChain instead.',
+    E_USER_DEPRECATED
+);
+
+/**
+ * @deprecated since 3.x, to be removed in 4.0.
+ */
+class AdapterChain extends \Sonata\Doctrine\Adapter\AdapterChain implements AdapterInterface
 {
-    private $adapters = [];
-
-    public function addAdapter(AdapterInterface $adapter): void
-    {
-        $this->adapters[] = $adapter;
-    }
-
-    public function getNormalizedIdentifier($model): ?string
-    {
-        foreach ($this->adapters as $adapter) {
-            $identifier = $adapter->getNormalizedIdentifier($model);
-
-            if ($identifier) {
-                return $identifier;
-            }
-        }
-
-        return null;
-    }
-
-    public function getUrlsafeIdentifier($model): ?string
-    {
-        foreach ($this->adapters as $adapter) {
-            $safeIdentifier = $adapter->getUrlsafeIdentifier($model);
-
-            if ($safeIdentifier) {
-                return $safeIdentifier;
-            }
-        }
-
-        return null;
-    }
 }