+++ tests/Fixtures/Entity/EntityManager.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Sonata\CoreBundle\Tests\Fixtures\Entity;
+
+use Sonata\CoreBundle\Model\BaseEntityManager;
+
+class EntityManager extends BaseEntityManager
+{
+}