+++ tests/Fixtures/Model/DocumentManager.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Sonata\CoreBundle\Tests\Fixtures\Model;
+
+use Sonata\CoreBundle\Model\BaseDocumentManager;
+
+/**
+ * @group legacy
+ */
+class DocumentManager extends BaseDocumentManager
+{
+}